requirements.txt
Django==1.6
django-tastypie==0.11.0
ipython==2.0.0